# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoHeal application monitoring and recovery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard aggregates
  - name: applications
    description: Monitored applications
  - name: services
    description: Monitored services
  - name: incidents
    description: Incidents and investigation timelines
  - name: recovery
    description: Safe recovery rules and actions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /v1/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Current database-backed dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /v1/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List registered applications
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Register an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /v1/applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get an application
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Application deleted
  /v1/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List monitored services
      parameters:
        - name: applicationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Register a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /v1/services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a monitored service
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a monitored service
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a monitored service
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Service deleted
  /v1/incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, INVESTIGATING, RECOVERING, RESOLVED, ESCALATED]
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
  /v1/incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get an incident with timeline and AI insight
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/incidents/{id}/analyze:
    post:
      operationId: analyzeIncident
      tags: [incidents]
      summary: Request a backend AI diagnosis
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: AI insight, or a clear not-connected response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIInsight"
  /v1/recovery/rules:
    get:
      operationId: listRecoveryRules
      tags: [recovery]
      summary: List safe recovery rules
      responses:
        "200":
          description: Recovery rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecoveryRule"
    post:
      operationId: createRecoveryRule
      tags: [recovery]
      summary: Create an allowlisted recovery rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryRuleInput"
      responses:
        "201":
          description: Recovery rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryRule"
  /v1/recovery/rules/{id}:
    patch:
      operationId: updateRecoveryRule
      tags: [recovery]
      summary: Update a recovery rule
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryRuleUpdate"
      responses:
        "200":
          description: Recovery rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryRule"
    delete:
      operationId: deleteRecoveryRule
      tags: [recovery]
      summary: Delete a recovery rule
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Recovery rule deleted
  /v1/recovery/actions:
    get:
      operationId: listRecoveryActions
      tags: [recovery]
      summary: List recovery actions
      responses:
        "200":
          description: Recovery actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecoveryAction"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DashboardSummary:
      type: object
      required:
        - totalApplications
        - healthyServices
        - activeIncidents
        - recoveryActions
        - failedHealthChecks
        - recentDeployments
        - monitoringConnected
        - monitoringMessage
      properties:
        totalApplications: { type: integer }
        healthyServices: { type: integer }
        activeIncidents: { type: integer }
        recoveryActions: { type: integer }
        failedHealthChecks: { type: integer }
        recentDeployments: { type: integer }
        monitoringConnected: { type: boolean }
        monitoringMessage: { type: string }
    Application:
      type: object
      required: [id, name, description, environment, baseUrl, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        environment: { type: string, enum: [development, staging, production] }
        baseUrl: { type: string, format: uri }
        status: { type: string, enum: [HEALTHY, DEGRADED, DOWN, NOT_CONNECTED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApplicationInput:
      type: object
      required: [name, environment, baseUrl]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        environment: { type: string, enum: [development, staging, production] }
        baseUrl: { type: string, format: uri }
    ApplicationUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        environment: { type: string, enum: [development, staging, production] }
        baseUrl: { type: string, format: uri }
    Service:
      type: object
      required: [id, applicationId, name, type, endpoint, healthEndpoint, expectedStatus, timeout, interval, status, consecutiveFailures, createdAt]
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        name: { type: string }
        type: { type: string }
        endpoint: { type: string, format: uri }
        healthEndpoint: { type: string, format: uri }
        expectedStatus: { type: integer }
        timeout: { type: integer }
        interval: { type: integer }
        status: { type: string, enum: [HEALTHY, DEGRADED, DOWN, NOT_CONNECTED] }
        consecutiveFailures: { type: integer }
        lastCheckedAt: { type: ["string", "null"], format: date-time }
        responseTime: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
    ServiceInput:
      type: object
      required: [applicationId, name, type, endpoint, healthEndpoint]
      properties:
        applicationId: { type: integer }
        name: { type: string, minLength: 1 }
        type: { type: string }
        endpoint: { type: string, format: uri }
        healthEndpoint: { type: string, format: uri }
        expectedStatus: { type: integer, default: 200 }
        timeout: { type: integer, default: 5000 }
        interval: { type: integer, default: 60 }
    ServiceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string }
        endpoint: { type: string, format: uri }
        healthEndpoint: { type: string, format: uri }
        expectedStatus: { type: integer }
        timeout: { type: integer }
        interval: { type: integer }
    Incident:
      type: object
      required: [id, applicationId, serviceId, type, severity, status, detectedAt, failureCount, description, recoveryAttempts]
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        serviceId: { type: integer }
        type: { type: string }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        status: { type: string, enum: [OPEN, INVESTIGATING, RECOVERING, RESOLVED, ESCALATED] }
        detectedAt: { type: string, format: date-time }
        resolvedAt: { type: ["string", "null"], format: date-time }
        failureCount: { type: integer }
        description: { type: string }
        recoveryAttempts: { type: integer }
    IncidentTimelineEvent:
      type: object
      required: [id, eventType, message, timestamp]
      properties:
        id: { type: integer }
        eventType: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
    IncidentDetail:
      allOf:
        - $ref: "#/components/schemas/Incident"
        - type: object
          required: [timeline, aiInsight]
          properties:
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/IncidentTimelineEvent"
            aiInsight:
              anyOf:
                - $ref: "#/components/schemas/AIInsight"
                - type: "null"
    AIInsight:
      type: object
      required: [status, diagnosis, rootCause, evidence, confidence, recommendations, prevention]
      properties:
        status: { type: string, enum: [AVAILABLE, NOT_CONNECTED, NO_DATA] }
        diagnosis: { type: string }
        rootCause: { type: string }
        evidence:
          type: array
          items: { type: string }
        confidence: { type: ["number", "null"] }
        recommendations:
          type: array
          items: { type: string }
        prevention:
          type: array
          items: { type: string }
    RecoveryRule:
      type: object
      required: [id, name, triggerType, threshold, actionType, enabled, cooldownSeconds, maxAttempts, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        triggerType: { type: string, enum: [CONSECUTIVE_FAILURES, HTTP_5XX, SERVICE_UNHEALTHY] }
        threshold: { type: integer }
        actionType: { type: string, enum: [RESTART_CONTAINER, START_CONTAINER, RETRY_HEALTH_CHECK] }
        target: { type: ["string", "null"], description: "Docker container name or ID for container actions" }
        enabled: { type: boolean }
        cooldownSeconds: { type: integer }
        maxAttempts: { type: integer }
        createdAt: { type: string, format: date-time }
    RecoveryRuleInput:
      type: object
      required: [name, triggerType, threshold, actionType, cooldownSeconds, maxAttempts]
      properties:
        name: { type: string, minLength: 1 }
        triggerType: { type: string, enum: [CONSECUTIVE_FAILURES, HTTP_5XX, SERVICE_UNHEALTHY] }
        threshold: { type: integer, minimum: 1 }
        actionType: { type: string, enum: [RESTART_CONTAINER, START_CONTAINER, RETRY_HEALTH_CHECK] }
        target: { type: string, description: "Docker container name or ID for container actions" }
        enabled: { type: boolean, default: true }
        cooldownSeconds: { type: integer, minimum: 0 }
        maxAttempts: { type: integer, minimum: 1 }
    RecoveryRuleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        threshold: { type: integer, minimum: 1 }
        actionType: { type: string, enum: [RESTART_CONTAINER, START_CONTAINER, RETRY_HEALTH_CHECK] }
        target: { type: string, description: "Docker container name or ID for container actions" }
        enabled: { type: boolean }
        cooldownSeconds: { type: integer, minimum: 0 }
        maxAttempts: { type: integer, minimum: 1 }
    RecoveryAction:
      type: object
      required: [id, actionType, status, startedAt]
      properties:
        id: { type: integer }
        incidentId: { type: ["integer", "null"] }
        actionType: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCESS, FAILED, BLOCKED] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        result: { type: ["string", "null"] }