# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AutoHeal application monitoring and recovery API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalApplications: number;
  healthyServices: number;
  activeIncidents: number;
  recoveryActions: number;
  failedHealthChecks: number;
  recentDeployments: number;
  monitoringConnected: boolean;
  monitoringMessage: string;
}