# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AutoHeal application monitoring and recovery API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}