# lib/api-zod/src/generated/types/incidentDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AutoHeal application monitoring and recovery API
 * OpenAPI spec version: 0.1.0
 */
import type { AIInsight } from './aIInsight';
import type { Incident } from './incident';
import type { IncidentTimelineEvent } from './incidentTimelineEvent';

export type IncidentDetail = Incident & ({
  timeline: IncidentTimelineEvent[];
  aiInsight: AIInsight | null;
});